# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ent.Question:
    properties:
      body:
        description: Body holds the value of the "body" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      order:
        description: Order holds the value of the "order" field.
        type: integer
    type: object
  ent.Questionnaire:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.QuestionnaireResponse:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      id:
        description: ID of the ent.
        type: string
    type: object
...
  types.BaseAnswer:
    properties:
      body:
        description: The answer body.
        example: 我這週心情還不錯！
        type: string
      question_id:
        description: |-
          The question this answer relates to, the question also needs to be in
          the same questionnaire as the response.
        type: string
    type: object
...
  types.BaseQuestion:
    properties:
      body:
        description: The question body
        example: 你這週的心情如何？
        type: string
    type: object
  types.BaseQuestionnaire:
    description: BaseQuestionnaire
    properties:
      name:
        description: The name of the questionnaire
        example: 問卷標題
        type: string
      questions:
        description: |-
          The initial questions in this questionnaire. This field may be empty
          and you can add questions later using post request to
          `quesionnaires/:id/new/question`.
        items:
          $ref: '#/definitions/types.BaseQuestion'
        type: array
    type: object
  types.BaseResponse:
    properties:
      answers:
        description: The answers to all questions in a questionnaire.
        items:
          $ref: '#/definitions/types.BaseAnswer'
        type: array
    type: object
...
  types.QuestionnaireDetails:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
      questions:
        items:
          $ref: '#/definitions/ent.Question'
        type: array
      responses:
        items:
          $ref: '#/definitions/types.Response'
        type: array
    type: object
...
  types.Response:
    properties:
      answers:
        description: The answers to all questions in a questionnaire.
        items:
          $ref: '#/definitions/types.BaseAnswer'
        type: array
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      user_id:
        description: The user ID of the user who submit the response.
        type: string
    type: object
  types.ResponseWithQuestionnaire:
    properties:
      answers:
        description: The answers to all questions in a questionnaire.
        items:
          $ref: '#/definitions/types.BaseAnswer'
        type: array
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      questionnaire:
        $ref: '#/definitions/ent.Questionnaire'
      user_id:
        description: The user ID of the user who submit the response.
        type: string
    type: object
...
host: health-statistic.dechnology.com.tw
info:
  contact: {}
  description: "This is the API documentation for 「健康資料公鏈」\n\n# Usage\n\n## Flow Diagram\n![flow_diagram](https://cdn.discordapp.com/attachments/874556062815100940/1132920083174408222/App-.drawio.png)\n\n##
    API Endpoints for Frontend Application\n\nFull descriptions of all endpoints are
    described below; however, more common endpoints for a client application are described
    here. For detailed types of params, body, or response, please refer to the full
    descriptiions below. \n\n### Server Health Check - `/health_check` [**GET**]\n\nThis
    endpoint checks if the server is alive and the database is connected. It returns
    a code of 200 if ture.\n\n---\n\n### Get Registration Questionnaire - `/questionnaires/registration`
    [**GET**]\n\nFor the sake of scalibility, we make the questionnaire system a general
    one. That is, we can later create more questionnaires in our application. Upon
    initialization of our database, the registration questionnaire is created and
    assigned a fixed UUID. Before the user register, we need to have them fill in
    the registration questionnaire. This questionniare can be fetch using this endpoint.
    Note that this questionnaire correspond to question 16 to 30 since the answers
    to question 1 to 15 are part of the user data.\n\n---\n\n### User Registration
...
    some answers are part of the user data like \"birth_year\", \"gender\", and so
...
paths:
...
  /questionnaires:
    get:
      description: |
        Get all questionnaires from the database. For each questionnaire, all questions and responses of this questionnaire are also included.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.QuestionnaireDetails'
            type: array
      summary: Get Questionnaires
      tags:
      - Questionnaire
    post:
      consumes:
      - application/json
      description: "Create a new questionnaire and optionally its questions.\n\n##
        Request Body\n\n- `name` *`string`* **Required**\n    The name of the questionnaire
        to be created.\n\n- `questions` *`array`* Optional\n    The initial questions
        in this questionnaire. This field may be empty and you can add questions later
        using post request to `quesionnaires/:id/new/question`.\n    \n    - `body`
        *`string`* **Required**\n        The body of one of the question in the questionnaire.\n
        \   \n    - `type` *`string`* **Required**\n        The type of the question.
        For now, we accept all strings but in the future this field might be an enum."
      parameters:
      - description: The questionnaire to be created.
        in: body
        name: questionnaire
        required: true
        schema:
          $ref: '#/definitions/types.BaseQuestionnaire'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Questionnaire'
      summary: Create Questionnaire
      tags:
      - Questionnaire
  /questionnaires/{id}:
    delete:
      description: Delete the questionnaire specified by the `id` path param.
      parameters:
      - description: The questionnaire's ID.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete Questionnaire
      tags:
      - Questionnaire
    get:
      description: |
        Get the questionnaire specified by the `id` path param. All questions and responses of the questionnaire are also included.
      parameters:
      - description: The questionnaire's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.QuestionnaireDetails'
      summary: Get Questionnaire
      tags:
      - Questionnaire
  /questionnaires/{id}/new/question:
    post:
      consumes:
      - application/json
      description: |-
        Create a new question for the questionnaire  by the `id` path param.

        ## Request Body

        - `body` *`string`* **Required**
            The body of one of the question in the questionnaire.

        - `type` *`string`* **Required**
            The type of the question. For now, we accept all strings but in the future this field might be an enum.
      parameters:
      - description: The questionnaire's ID.
        in: path
        name: id
        required: true
        type: string
      - description: The question to be created.
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/types.BaseQuestion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Question'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ent.Question'
      summary: Create Question
      tags:
      - Questionnaire
  /questionnaires/{id}/new/response:
    post:
      consumes:
      - application/json
      description: "Create a new response for the questionnaire specified by `id`
        path param.\n\n## Request Body\n\n- `user_id` *`string`* **Required**\n    The
        user auth0 ID. This is the user who submits this response.\n\n- `answers`
        *`array`* **Required**\n    The answers to the questions in the questionnaire
        submitted by the user above.Note that the length of the answers array must
        equal to the number of questions in the given questionnaire. Also, all the
        `question_id`s must match the questions in the questionnaire.\n \n    - `question_id`
        *`string`* **Required**\n    The question id of a question in the questionnaire
        which this answer correspond to.\n\n    - `body` *`string`* **Required**\n
        \   The body of the answer.\n"
      parameters:
      - description: The questionnaire's ID.
        in: path
        name: id
        required: true
        type: string
      - description: The response to be created.
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/types.BaseResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.QuestionnaireResponse'
      summary: Create Response
      tags:
      - Questionnaire
...
  /responses/{id}:
    delete:
      description: Delete the response specified by the `id` path param.
      parameters:
      - description: The response's ID.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete Response
      tags:
      - Response
    get:
      description: |
        Get the response specified by the `id` path param. **This will also includes the corresponding questionnaire.**
      parameters:
      - description: The response's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ResponseWithQuestionnaire'
      summary: Get Response
      tags:
      - Response
...
schemes:
...
swagger: "2.0"